refactor: Post.swift split up into several classes
