+++ Imagine/Model/PostClasses/Post.swift
@@ -10,95 +10,9 @@ import Foundation
 import Firebase
 import FirebaseFirestore
 
-// This enum declares the type of the post
-enum PostType {
-    case picture
-    case link
-    case thought
-    case repost
-    case translation
-    case youTubeVideo
-    case GIF
-    case multiPicture
-    case panorama
-    case singleTopic
-    case topTopicCell
-    case nothingPostedYet
-}
-
-// This enum declares how a post was marked or reported
-enum ReportType {
-    case normal
-    case spoiler
-    case satire
-    case opinion
-    case sensationalism
-    case circlejerk
-    case pretentious
-    case edited
-    case ignorant
-    case misinformation
-    case misleading
-}
-
-class Votes {
-    var thanks = 0
-    var wow = 0
-    var ha = 0
-    var nice = 0
-}
-
-enum MusicType {
-    case track
-    case album
-}
-
-class Music {
-    var type: MusicType
-    var name: String
-    var artist: String
-    var releaseDate: Date?
-    var artistImageURL: String?
-    var musicImageURL: String
-    var songwhipURL: String
-    
-    ///Playlist init
-    init(type:MusicType, name: String, artist: String, musicImageURL: String, songwhipURL: String) {
-        self.type = type
-        self.name = name
-        self.artist = artist
-        self.musicImageURL = musicImageURL
-        self.songwhipURL = songwhipURL
-    }
-    
-    init(type: MusicType, name: String, artist: String, releaseDate: Date, artistImageURL: String, musicImageURL: String, songwhipURL: String) {
-        self.type = type
-        self.name = name
-        self.artist = artist
-        self.releaseDate = releaseDate
-        self.artistImageURL = artistImageURL
-        self.musicImageURL = musicImageURL
-        self.songwhipURL = songwhipURL
-    }
-}
-
-class Link {
-    var imageURL: String?
-    var link: String
-    var shortURL: String
-    var linkTitle: String
-    var linkDescription: String
-    
-    init(link: String, title: String, description: String, shortURL: String, imageURL: String?) {
-        self.link = link
-        self.linkTitle = title
-        self.linkDescription = description
-        self.shortURL = shortURL
-        self.imageURL = imageURL
-    }
-}
-
 class Post {
+    
+    //MARK:- Variables
     var title = ""
     var imageURL = ""
     var imageURLs: [String]?
@@ -138,6 +52,7 @@ class Post {
     let db = Firestore.firestore()
     
     
+    //MARK: Get Repost
     func getRepost(returnRepost: @escaping (Post) -> Void) {
         if let repostID = repostDocumentID {
             var collectionRef: CollectionReference!
@@ -171,6 +86,8 @@ class Post {
             })
         }
     }
+    
+    //MARK:- Get User
     
     func getUser(isAFriend: Bool) {
         
@@ -233,72 +150,5 @@ class Post {
                 }
             }
         })
-    }
-    
-//    func getCommentCount() {
-//
-//        if self.documentID != "" {
-//            let commentRef = db.collection("Comments").document(self.documentID).collection("threads")
-//
-//            commentRef.getDocuments { (snap, err) in
-//                if let err = err {
-//                    print("Wir haben einen Error beim User: \(err.localizedDescription)")
-//                }
-//                if let snapshot = snap {
-//                    self.commentCount = snapshot.count
-//                }
-//            }
-//        }
-//    }
-}
-
-
-
-public class User {
-    public var displayName = ""
-    public var imageURL = ""
-    public var userUID = ""
-    public var image = UIImage(named: "default-user")
-    public var blocked: [String]?
-    public var statusQuote = ""
-    
-    //Social Media Links
-    public var instagramLink: String?
-    public var instagramDescription: String?
-    public var patreonLink: String?
-    public var patreonDescription: String?
-    public var youTubeLink: String?
-    public var youTubeDescription: String?
-    public var twitterLink: String?
-    public var twitterDescription: String?
-    public var songwhipLink: String?
-    public var songwhipDescription: String?
-    
-    //location
-    public var locationName: String?
-    public var locationIsPublic = false
-    
-    let db = Firestore.firestore()
-    
-    func getBadges(returnBadges: @escaping ([String]) -> Void) {
-        
-        if userUID != "" {
-            let ref = db.collection("Users").document(userUID)
-            ref.getDocument { (snap, err) in
-                if let error = err {
-                    print("We have an error: \(error.localizedDescription)")
-                } else {
-                    if let snap = snap {
-                        if let data = snap.data() {
-                            if let badges = data["badges"] as? [String] {
-                                returnBadges(badges)
-                            }
-                        }
-                    }
-                }
-            }
-        } else {
-            print("Got no UID for badges")
-        }
     }
 }
