+++ Imagine/Model/PostClasses/ReportType.swift
@@ -1,0 +1,24 @@
+//
+//  ReportType.swift
+//  Imagine
+//
+//  Created by Malte Schoppe on 27.02.21.
+//  Copyright © 2021 Malte Schoppe. All rights reserved.
+//
+
+import Foundation
+
+/// An enum that declares how a post was highlighted or reported
+enum ReportType {
+    case normal
+    case spoiler
+    case satire
+    case opinion
+    case sensationalism
+    case circlejerk
+    case pretentious
+    case edited
+    case ignorant
+    case misinformation
+    case misleading
+}
