+++ Imagine/Model/PostClasses/Votes.swift
@@ -1,0 +1,16 @@
+//
+//  Votes.swift
+//  Imagine
+//
+//  Created by Malte Schoppe on 27.02.21.
+//  Copyright © 2021 Malte Schoppe. All rights reserved.
+//
+
+import Foundation
+
+class Votes {
+    var thanks = 0
+    var wow = 0
+    var ha = 0
+    var nice = 0
+}
